refactor(sed): read-safe generic builder replaces 14 bespoke wrappers (python)
sed joins the generic_bind builders: the stream-to-stdout path works on
every read-capable backend and only -i needs a write op, failing with
'sed: -i not supported on this backend'. Deletes the bespoke wrappers,
make_sed, and the dead databricks _helpers; databricks -i now works
(the backend has a write op, the old refusal was a wrapper limitation).
