--- python/mirage/commands/builtin/hf_buckets/sed.py
@@ -1,28 +1,0 @@
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-
-from mirage.accessor._hf import HF_RESOURCES
-from mirage.commands.builtin.generic.sed_command import make_sed
-from mirage.core.hf_buckets.glob import resolve_glob
-from mirage.core.hf_buckets.read import read_bytes
-from mirage.core.hf_buckets.stat import stat as _stat
-from mirage.core.hf_buckets.write import write_bytes
-
-sed = make_sed(
-    stat_fn=_stat,
-    resource=HF_RESOURCES,
-    glob_fn=resolve_glob,
-    make_read=lambda accessor, index, paths: read_bytes,
-    write_bytes=write_bytes,
-)
