+++ python/mirage/commands/builtin/ram/__init__.py
@@ -15,7 +15,6 @@
 from mirage.commands.builtin.filetype_factory import make_filetype_commands
 from mirage.commands.builtin.generic_bind import (CommandIO,
                                                   make_generic_commands)
-from mirage.commands.builtin.ram.sed import sed
 from mirage.core.ram.constants import SCOPE_ERROR
 from mirage.core.ram.copy import copy as _copy
 from mirage.core.ram.du import du as _du
@@ -59,5 +58,4 @@ _RAM_CMD_OPS = CommandIO(
 COMMANDS = [
     *make_filetype_commands("ram", _ft_resolve_glob, _ft_read),
     *make_generic_commands("ram", _RAM_CMD_OPS),
-    sed,
 ]
