+++ spec/python/general/sed.json
@@ -8,20 +8,33 @@
       "chroma",
       "databricks_volume",
       "dify",
+      "discord",
       "disk",
+      "email",
+      "gdocs",
       "gdrive",
       "github",
+      "github_ci",
+      "gmail",
+      "gsheets",
+      "gslides",
       "hf_buckets",
-      "hf_datasets",
-      "hf_models",
-      "hf_spaces",
+      "lancedb",
+      "langfuse",
+      "linear",
+      "mongodb",
       "nextcloud",
+      "notion",
       "onedrive",
+      "postgres",
+      "qdrant",
       "ram",
       "redis",
       "s3",
       "sharepoint",
-      "ssh"
+      "slack",
+      "ssh",
+      "trello"
     ]
   },
   "description": null,
